requirements.txt
# Install PyTorch with CUDA support as needed.
# For example, if using CUDA 12.6, run:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu126

openai-whisper
sounddevice>=0.4.6
numpy>=1.24
scipy>=1.10.0
webrtcvad
requests>=2.31.0
pyyaml>=6.0.1
snac
